Lab 9 - Airflow/requirements.txt
apache-airflow==2.6.3
pendulum==2.1.2
Flask-Session==0.5.0
scikit-learn
pandas
numpy
lightgbm
joblib
gradio
connexion[swagger-ui]==2.14.1
